submit java-oop-ru/interfaces

+++ java-oop-ru/interfaces/src/main/java/exercise/Cottage.java
@@ -10,7 +10,7 @@ public class Cottage implements Home {
         this.floorCount = floorCount;
     }
 
-    public int getArea() {
+    public double getArea() {
         return this.area;
     }
