+++ java-oop-ru/interfaces/src/main/java/exercise/Home.java
@@ -2,7 +2,7 @@ package exercise;
 
 // BEGIN
 public interface Home {
-    int getArea();
+    double getArea();
     int compareTo(Home another);
 }
 // END
